Set up USWDS theme files, reorganize/refactor components, resolve all linter warnings
