+++ Procfile.dev
@@ -1,2 +1,3 @@
 webpack: react-scripts start
 web: PROXY_PORT=4000 nodemon -V --ignore ./src --ignore ./test ./bin/www
+sass: sass --load-path=node_modules/uswds/dist/scss --style=compressed --watch theme/styles.scss build/theme/styles.css
