+++ app.js
@@ -29,9 +29,15 @@ app.use(
 app.use(passport.initialize());
 app.use(passport.session());
 
-app.use('/libraries/uswds', express.static(path.join(__dirname, 'node_modules/uswds/dist')));
 app.use(express.static(path.join(__dirname, 'build'), { index: false }));
 app.use(express.static(path.join(__dirname, 'public'), { index: false }));
+
+app.use('/theme', express.static(path.join(__dirname, 'node_modules/uswds/dist')));
+if (process.env.NODE_ENV !== 'production') {
+  // for css sourcemap debugging only
+  app.use('/theme', express.static(path.join(__dirname, 'theme')));
+  app.use('/node_modules/uswds/dist', express.static(path.join(__dirname, 'node_modules/uswds/dist')));
+}
 app.use('/', require('./routes'));
 
 app.get('/*', isAuthenticated, (req, res) => {
