+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "^7.1.2",
     "bcrypt": "^5.0.0",
     "bootstrap": "^4.5.2",
+    "classnames": "^2.2.6",
     "cookie-parser": "^1.4.5",
     "debug": "^4.1.1",
     "ejs": "^3.1.5",
@@ -19,6 +20,7 @@
     "http-proxy-middleware": "^1.0.5",
     "http-status-codes": "^1.4.0",
     "lodash": "^4.17.20",
+    "luxon": "^1.25.0",
     "mocha": "^8.1.1",
     "morgan": "^1.10.0",
     "nodemon": "^2.0.4",
@@ -32,6 +34,8 @@
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
+    "sass": "^1.28.0",
+    "sass-loader": "^10.0.5",
     "sequelize": "^6.3.3",
     "sequelize-auto": "^0.5.3",
     "sequelize-cli": "^6.2.0",
@@ -47,7 +51,8 @@
     "start": "react-scripts start",
     "start:dev": "nf start -j Procfile.dev",
     "start:prod": "nf start",
-    "build": "react-scripts build",
+    "build": "react-scripts build && yarn sass",
+    "sass": "sass --load-path=node_modules/uswds/dist/scss --no-source-map --style=compressed theme/styles.scss build/theme/styles.css",
     "test": "yarn prettier --check . && yarn lint && yarn test:server && yarn test:client --watchAll=false",
     "test:client": "react-scripts test",
     "test:server": "mocha --recursive",
