+++ public/index.html
@@ -22,7 +22,7 @@
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
     <!-- USWDS CSS -->
-    <link rel="stylesheet" href="/libraries/uswds/css/uswds.min.css" />
+    <link rel="stylesheet" href="/theme/styles.css" />
     <title>React App</title>
   </head>
   <body>
