--- src/App.css
@@ -1,34 +1,0 @@
-.emsContainer {
-    background-color: #E5E5E5;
-    height: 90%;
-    width: 96%;
-    margin: auto;
-    text-align: justify;
-    border: 1px solid #beb8b8;
-  }
-.bedContent {
-  padding: 10% 10% 10% 10%;
-}
-.bedComponentContainer {
-  padding: 0% 10% 7% 3%;
-  border-bottom: 0.05px solid black;
-}
-a {
-  text-decoration: none;
-}
-
-.title {
-  text-align: center;
-}
-label {
-  margin: 0% 0% 0% 3%;
-}
-input {
-  margin: 0% 0% 5% 9.5%;
-  position: relative;
-  left: -130px;
-  top: 21px;
-}
-.emsLink{
-  margin: 10%;
-}
