+++ src/EMS/BedInfo.js
