+++ src/EMS/index.js
@@ -1,0 +1,3 @@
+import EMS from './EMS';
+
+export default EMS;
