+++ src/ER/RingDown.js
@@ -1,11 +1,11 @@
-import React, { useState } from 'react';
+import React from 'react';
 
 const RingDown = () => {
   return (
     <div>
       <h1> Answer Ringdown </h1>
     </div>
-  )
+  );
 };
 
 export default RingDown;
