--- src/containers/components/NavBar.js
@@ -1,10 +1,0 @@
-import React from 'react'
-
-export default function NavBar ({setSelectedTab}) { 
-    return ( 
-        <div> 
-            <div onClick={() => setSelectedTab("beds")}>Beds</div>
-            <div onClick={() => setSelectedTab("ringDown")}>RingDown</div>
-        </div>
-    )
-}
