--- src/css/nav.css
@@ -1,104 +1,0 @@
-.header h2{
-  background: #1A4480;
-  width: 320px;
-  height: 67px;
-  font-family: sans-serif;  
-}
-#appName{
-  color: white;
-  height: 67px;
-  text-align: left;
-  line-height: 77px;
-  font-size: 16px;
-  padding-left: 10px;
-}
-#appTimeDisplay{
-  position: relative;
-  top: -6px;
-  color: white;
-  height: 47px;
-  text-align: left;
-  line-height: 6px;
-  font-size: 32px;
-  padding-left: 10px;
-}
-.nav {
-  background-color: white;
-  /* border-bottom: 0.5px solid grey; */
-  /* box-shadow: 0px 1px 1px grey; */
-  width: 320px;
-  height: 64px;
-  position: relative;
-  left: 0px;
-  top: -71px;
-  padding-left: 10px;
-}
-h2 {
-  text-align: center;
-  position: relative;
-  top: 24px;
-  color: grey;
-}
-.beds {
-  position: relative;
-  top: -52.73px;
-  left: 158px;
-  font-size: 16px;
-  border: none;
-  color: gray;
-  height: 49px;
-  border-radius: 20px 0px 0px 20px;
-
-}
-.ringDown {
-  position: relative;
-  top: -0.73px;
-  left: -10px;
-  font-size: 16px;
-  border: none;
-  color: gray;
-  border-radius: 0px 20px 20px 0px;
-  height: 52px;
-  
-}
-.ringDownTab .ringDown{
-  width: 162px;
-  text-align: center;
-  border-radius: none;
-  background: #F0F0F0;  
-}
-.ringDownTabselected .ringDown{
-  width: 162px;
-  text-align: center;
-  color: #0050D8;
-  background-color: white;
-}
-
-.bedTab .beds {
-  width: 153px;
-  text-align: center;
-  border-radius: none;
-  background: #F0F0F0; 
-}
-.bedTabSelected .beds {
-  /* width: 162px; */
-  text-align: center;
-  color: #0050D8;
-  background-color: white;
-}
-/* button {
-  
-  border: none;
-  border-bottom: #0050D8;
-  background-color: white;
-} */
-
-
-
-/* width: 50%;
-position: relative;
-left: 615px;
-top: 1px;
-text-align: center;
-color: grey; */
-
