--- src/css/ringDown.css
@@ -1,416 +1,0 @@
-#legend-header{
-  border-bottom: 1px solid black;
- margin: 0px 0px 40px 0px;
- background-color: rgb(243, 243, 243);
-color: black;
-text-align: start;
-}
-.header {
-  position: relative;
-  top: -37px;
-}
-.RingDownInfo {
-  position: relative;
-  left: 2em;
-  width: 320px;
-  height: 132px;
-  font-size: 16px;
-  color: #71767A;
-  /* border-bottom: 1px solid #DCDEE0; */
-}
-#unitLabel {
-  position: relative;
-  left: -30px;
-  top: -57px;
-  width: 272px;
-  height: 24px;
-  text-align: left;
-}
- #unit {
-  position: relative;
-  left: -82px;
-  top: 34px;
-  width: 140px;
-  height: 36px; 
-}
-#incidentLabel {
-  position: relative;
-  left: -270px;
-  top: 29px;
-  width: 272px;
-  height: 24px;
-  text-align: left;
-  margin: 7px;
-}
-#incidentId {
-  position: relative;
-  left: 217px;
-  top: -23px;
-  width: 140px;
-  height: 36px; 
-}
-.title {
-  position: relative;
-  left: -1px;
-  top: -25px;
-  text-align: left;
-  font-size: 16px;
-  color: #3D4551;
-  border-bottom: 1px solid black;
-  padding-bottom: 10px;
-  width: 320px;
-  padding-left: 10px;
-}
-/* #ageLabel {
-  position: relative;
-  left: -27px;
-  top: -64px;
-  font-size: 16px;
-}
-#age {
-  position: relative;
-  left: -123px;
-  top: 33px;
-  height: 36px;
-  width: 64px;
-  border-radius: 4px;
-}
-#maleLabel {
-  position: relative;
-  top: 23px;
-  left: 1px;
-  font-size: 16px;
-}
-#male {
-  position: relative;
-  top: 5px;
-  left: -98px;
-  height: 22px;
-  width: 22px;
-}
-#femaleLabel {
-  position: relative;
-  top: 62px;
-  left: -146px;
-  font-size: 16px;
-}
-#female {
-  position: relative;
-  top: 5px;
-  left: -117px;
-  height: 22px;
-  width: 22px;
-}
-#non-binaryLabel {
-  position: relative;
-  top: 97px;
-  left: -228px;
-  font-size: 16px;
-}
-#non-binary {
-  position: relative;
-  top: -30px;
-  left: 177px;
-  height: 22px;
-  width: 22px;
-}
-.genderLabels {
-  position: relative;
-  top: -116px;
-}
-#code2Label {
-  position: relative;
-  top: 97px;
-  left: -1px;
-  font-size: 16px;
-}
-#code2 {
-  position: relative;
-  top: 4px;
-  left: -146px;
-  height: 22px;
-  width: 22px;
-}
-#code3Label {
-  position: relative;
-  top: 134px;
-  left: -170px;
-  font-size: 16px;
-}
-#code3 {
-  position: relative;
-  top: 4px;
-  left: -146px;
-  height: 22px;
-  width: 22px;
-}
-.codeLabels {
-  position: relative;
-  top: -182px;
-}
-#complaintLabel {
-  position: relative;
-  top: 158px;
-  left: -5px;
-  font-size: 16px;
-}
-#complaint {
-  position: relative;
-  top: 4px;
-  left: -5px;
-  height: 136px;
-  width: 268px;
-}
-#stableLabel {
-  position: relative;
-  top: 213px;
-  left: -1px;
-  font-size: 16px;
-}
-#stable {
-  position: relative;
-  top: 4px;
-  left: -161px;
-  height: 22px;
-  width: 22px;
-}
-#unstableLabel {
-  position: relative;
-  top: 248px;
-  left: -175px;
-  font-size: 16px;
-}
-#unstable {
-  position: relative;
-  top: -33px;
-  left: 127px;
-  height: 22px;
-  width: 22px;
-}
-.patientInfo {
-  display: none;
-}
-#patientDisplay {
-  position: relative;
-  top: -87px;
-  left: 301px;
-  border: none;
-  background-color: white;
-  font-size: 16px;
-}
-#vitalsDisplay {
-  position: relative;
-  top: -115px;
-  left: 301px;
-  background-color: white;
-  border: none;
-  font-size: 16px;
-}
-#vitalsDisplay2 {
-  position: relative;
-  top: -96px;
-  left: 303px;
-  background-color: white;
-  border: none;
-  font-size: 16px;
-}
-.vitalsInfo {
-  display: none;
-}
-#vitals {
-  position: relative;
-  top: -74px;
-
-  text-align: left;
-}
-#vitals2 {
-  position: relative;
-  top: -51px;
-   
-}
-#vitalsContainer {
-  position: relative;
-  top: 264px;
-}
-#bpLabel{
-  position: relative;
-  top: -70px;
-  left: -209px;
-  font-size: 16px;
-  margin: 194px;
-}
-#bp {
-  position: relative;
-  top: 4px;
-  left: 164px;
-  width: 168px;
-  height: 36px;
-}
-#saveButton{
-  position: relative;
-  top: 845px;
-  left: 25px;
-  background-color: #005EA2;
-  color: white;
-  width: 272px;
-  height: 44px;
-}
-#pulseLabel {
-  position: relative;
-  top: -67px;
-  left: -43px;
-  margin: 29px;
-}
-#pulse {
-  position: relative;
-  top: 32px;
-  left: -129px;
-  width: 112px;
-  height: 36px;
-}
-#respiratoryLabel {
-  position: relative;
-  top: -29px;
-  left: -189px;
-  margin: 179px;
-}
-#respiratory {
-  position: relative;
-  top: 4px;
-  left: 149px;
-  width: 168px;
-  height: 36px;
-}
-#spo2Label {
-  position: relative;
-  top: -29px;
-  left: -201px;
-  margin: 190px;
-}
-#spo2 {
-  position: relative;
-  top: 4px;
-  left: 159px;
-  width: 88px;
-  height: 36px;
-}
-#RALabel {
-  position: relative;
-  top: -18px;
-  left: 69px;
-  font-size: 16px;
-}
-#RA {
-  position: relative;
-  top: 5px;
-  left: -121px;
-  height: 22px;
-  width: 22px;
-}
-#O2Label {
-  position: relative;
-  top: 25px;
-  left: -92px;
-  font-size: 16px;
-}
-#O2 {
-  position: relative;
-  top: 4px;
-  left: -121px;
-  height: 22px;
-  width: 22px;
-}
-#tempLabel {
-  position: relative;
-  top: 53px;
-  left: -201px;
-  margin: 190px;
-}
-#temp {
-  position: relative;
-  top: 4px;
-  left: 159px;
-  width: 88px;
-  height: 36px;
-}
-#notes {
-  position: relative;
-  top: -100px;
-}
-#notes2 {
-  position: relative;
-  top: 356bgpx;
-}
-.notesInfo {
-  display: none;
-}
-#notesDisplay {
-  position: relative;
-  top: -147px;
-  left: 300px;
-  background-color: white;
-  border: none;
-  font-size: 16px;
-}
- .infoAge {
-  height: 100px;
-}  
-.positionDown{
-  position: relative;
-  top: 93px;
-}
-#button-location{
-  position: relative;
-  top: 93px;
-} */
-
-.usa-radio__label{
-  border-radius: 40%;
-  color: black;
-}
-.personalInfo{
-  position: relative;
-  top: 5em;
-}
-
-.usa-accordion{
-  position: relative;
-  top: 5em;
-}
-
-.usa-accordion__content{
-  position: relative;
-  left: 0em;
-  top: 2em;
-}
-.additionalNotes{
-  position: relative;
-  top: 5em;
-}
-.usa-legend{
-  background-color: rgb(243, 243, 243);
-  border-bottom: 1px solid black;
-  color: black;
-  height: 44px;
-  font-size: 25px;
-}
-.usa-header{
-  border-bottom: 2px solid black;
-  background-color: rgb(243, 243, 243);
-  color: black;
-  position: relative;
-  top: 3em;
-  left: -1em;
-}
-.select{
-  position: relative;
-  top: 8em;
-}
-.clear{
-  position: relative;
-  top: 9em;
-}
-.usa-character-count{
-  position: relative;
-  top: 4em;
-}
