--- src/nav.css
@@ -1,65 +1,0 @@
-.nav{
-    background-color: white;
-    border-bottom: 0.5px solid grey;
-    box-shadow: 0px 0px 2px grey;
-    width: 100%;
-    position: relative;
-    left: 0px;
-    top: -20px;
-}
-h2 {
-    text-align: center;
-    position: relative;
-    top: 24px;
-    color: grey;
-}
-.bedTab {
-    width: 50%;
-    text-align: center;
-    position: relative;
-    top: 44px;
-  }
-  .bedTabSelected {
-    color: blue;
-    width: 50%;
-    text-align: center;
-    position: relative;
-    top: 44px;
-    border-bottom: 1px solid blue;
-  }
-  .beds {
-    font-size: 36px;
-    
-  }
-  .ringDown {
-    font-size: 36px;
-  }
-  .ringDownTab {
-    width: 50%;
-    position: relative;
-    left: 615px;
-    top: 1px;
-    text-align: center;
-  }
-  .ringDownTabselected {
-    color: blue;
-    width: 50%;
-    position: relative;
-    left: 615px;
-    top: 1px;
-    text-align: center;
-    border-bottom: 1px solid blue;
-  }
-  a:link {
-    color: black;
-  }
-  a:visited {
-    color: black;
-  }
-/* .container {
-    background-color: #E5E5E5;
-    height: 90%;
-    width: 96%;
-    margin: auto;
-    text-align: justify;
-  } */
