+++ views/layout.ejs
@@ -6,7 +6,7 @@
     <meta name="description" content="">
     <title>BATS</title>
     <!-- USWDS CSS -->
-    <link rel="stylesheet" href="/libraries/uswds/css/uswds.min.css" />
+    <link rel="stylesheet" href="/theme/styles.css" />
   </head>
   <body>
     <%- body %>
